Improve diff
pyvm diff is based on heuristics and there is a test
to determine whether the heuristics failed and then
try some expensive methods to generate the diff.

The change is: for "small" diffs we are more eager
to do the expensive methods.

This, for one, improves significantly the diffs we
see in gitex, which so far are mostly pretty small
and as a result are now more readable and easy to
figure out.

diffstudy shows an overall improvement. Results for
running diffstudy on the official git's repository:

Before:
 13606 diffs studied (history of 1151 files over 20483 commits)
 739 diffs differ
 From them:
  61 diffs differ big time
  432 diffs are better!
 External diff 194332 insertions (+), 115833 deletions (-), patchlines 607014
 Custom   diff 195540 insertions (+), 117041 deletions (-), patchlines 607808
 0.13% bigger patches
 External diff time: 103.32 seconds
 Custom   diff time: 48.88 seconds

After:
 13606 diffs studied (history of 1151 files over 20483 commits)
 710 diffs differ
 From them:
  49 diffs differ big time
  442 diffs are better!
 External diff 194332 insertions (+), 115833 deletions (-), patchlines 607014
 Custom   diff 195313 insertions (+), 116814 deletions (-), patchlines 607515
 0.08% bigger patches
 External diff time: 101.35 seconds
 Custom   diff time: 54.44 seconds

Also:
 - fixes in "\ No newline at end of file" things
 - fixed diffstudy which was broken all this time
 - implement `git checkin msgfile` which was then
   used to write this commit message here!