Subtitles (.srt) and sync fixes
External subtitle files can be used by avplay as:

   avplay movie.avi subtitles.srt
   avplay movie.avi -srt

With the second form avplay will look for a file `movie.srt`.

Note that the srt files should be in utf-8 but this is
not usually the case. In other words if you try to use
non-English subtitles and you get garbage the srt file
probably needs transcoding. For example, for greek subs
try:

   pyvm ucat 8859-7 movie.srt > movie2.srt

Also, a serious redesign of audio/video synchronization
and better resyncing after a seek. A step closer to
the right way to use libav*.