git auto stash II
The auto-stash feature from two commits ago works
but changes were not detected by 'git diff' because
the timestamp of the exported stashed files was
the same as the timestamp in the index cache.

This has been fixed by first checking out the
official branch tree and then doing a second
checkout of the stashed tree, waiting for some
time and touching the newly exported files.