linuxfb: use keyboard in RAW mode
The linux keyboard was previously used in MEDIUMRAW mode like SDL
but it seems that in MEDIUMRAW we don't have full control of all
the keys (for example, the function keys F1-F9 don't seem to work).

The keyboard handling code has been redesigned for the linuxfb
backend. Most of the work is now done at the python level.

Currently the code assumes a QWERTY AT keyboard. Anything else
will fail until we implement variable keyboard mappings.

A good source for this stuff is xserver/hw/kdrive/linux/keyboard.c