refactor aes
Initially AES worked only in CBC mode and had a fast assembly
implementation. Then, CFB mode was added basically to be able
to use protected secret keys in PGP.  The addition of the CFB
mode was a quick hack in order to move on with PGP.  Now this
has been refactored to cleaner code plus checks that CFB/CBC
work only in one-pass mode.