pyvm public key
The public key of pyvm has been recreated to get a subkey
that can be used for encryption. The key ID of the primary
key is the same (E1A88C45) so the git pullpgp scripts will
continue to work with the old key.

In order to import the new key, please edit the file

	pyvm/user/PGP/PUBKEYS.PGP

find the line that begins with E1A88C45 and delete it.
Then run

	pyvm gpg import pyvm-pubkey.pgp

to import the new key and feel free to send encrypted
messages, postcards and quotes with

	pyvm gpg encrypt <filename> E1A88C45

:)